Criacao do modulo de consultas

+++ backend/veterinaria/src/main/java/br/com/softvet/veterinaria/model/repository/ConsultasRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.softvet.veterinaria.model.entity.Consulta;
 
-public interface ConsultasRepository extends JpaRepository<Consulta, Integer> {
+public interface ConsultasRepository extends JpaRepository<Consulta, Long> {
 
 }
